face_recognition_server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
gunicorn==21.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
